Align implementation of block-attributes and block-info with with litvis
