+++ src/lib/block-attributes/index.ts
@@ -1,7 +1,4 @@
-export interface BlockAttributes {
-  [key: string]: any;
-}
-
-export { default as normalizeBlockAttributes } from "./normalize";
-export { default as parseBlockAttributes } from "./parse";
-export { default as stringifyBlockAttributes } from "./stringify";
+export * from "./normalizeBlockAttributes";
+export * from "./parseBlockAttributes";
+export * from "./stringifyBlockAttributes";
+export * from "./types";
