+++ src/lib/block-attributes/types.ts
@@ -1,0 +1,3 @@
+export interface BlockAttributes {
+  [key: string]: any;
+}
