+++ src/lib/block-info/index.ts
@@ -1,13 +1,3 @@
-import { BlockAttributes } from "../block-attributes";
-
-export interface BlockInfo {
-  attributes: BlockAttributes;
-  derivedAttributes?: BlockAttributes;
-  language: string;
-}
-
-export { default as normalizeBlockInfo } from "./normalize";
-export { default as parseBlockInfo } from "./parse";
-
-export const extractCommandFromBlockInfo = (info: BlockInfo) =>
-  info.attributes["cmd"] === true ? info.language : info.attributes["cmd"];
+export * from "./normalizeBlockInfo";
+export * from "./parseBlockInfo";
+export * from "./types";
