+++ src/lib/block-info/types.ts
@@ -1,0 +1,7 @@
+import { BlockAttributes } from "../block-attributes";
+
+export interface BlockInfo {
+  attributes: BlockAttributes;
+  derivedAttributes?: BlockAttributes;
+  language: string;
+}
